GG-37761 Only idempotent queries can be retried (wip).
